- Data Layer Implemented;

+++ domain/src/main/java/com/example/domain/repository/AppRepository.kt
@@ -160,8 +160,8 @@ interface AppRepository {
         position: String,
         weight: Int,
         working_leg: String,
-        lat: Int,
-        lon: Int,
+        lat: Double,
+        lon: Double,
         place_name: String,
     ): EditMyProfileResultEntity
